requirements.txt
# Feel free to add your Python dependencies below. For formatting guidelines, see:
# https://pip.pypa.io/en/latest/reference/requirements-file-format/

apify ~= 2.0.0
selenium
pandas
google-auth
google-auth-oauthlib
google-api-python-client
